change CUDA test

+++ test/test_cuda.py
@@ -805,12 +805,16 @@ class TestCuda(TestCase):
 
     def test_specify_improper_device_name(self):
         import os
-        with self.assertRaisesRegex(RuntimeError, "Expected one of cpu"):
-            fname = "tempfile.pt"
-            torch.save([torch.nn.Parameter(torch.randn(10, 10))], fname,
-                       _use_new_zipfile_serialization=True)
-            torch.load(fname, 'cuda0')
-        os.remove(fname)
+        fname = "tempfile.pt"
+        try:
+            with self.assertRaisesRegex(RuntimeError, "Expected one of cpu"):
+
+                torch.save([torch.nn.Parameter(torch.randn(10, 10))], fname,
+                           _use_new_zipfile_serialization=True)
+                torch.load(fname, 'cuda0')
+        finally:
+            if os.path.exists(fname):
+                os.remove(fname)
 
     def test_get_device_index(self):
         from torch.cuda._utils import _get_device_index
